pyproject.toml
[tool.poetry]
name = "talana-api"
version = "0.1.0"
description = "REST API test"
authors = ["Hugo Silva Álvarez <hugofer93@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
python-decouple = "~3.4"
Django = "~3.2"
djangorestframework = "~3.12"
drf-ujson2 = "~1.6"
celery = "~5.1"
numpy = "~1.20"

[tool.poetry.dev-dependencies]
pylint = "~2.8"
pycodestyle = "~2.7"
django-silk = "~4.1"

[build-system]
